# optimize TBasketKey unpacking