--- a/src/utils.jl
+++ b/src/utils.jl
@@ -60,6 +60,7 @@ function JaggType(f, branch, leaf)
     # https://github.com/scikit-hep/uproot3/blob/54f5151fb7c686c3a161fbe44b9f299e482f346b/uproot3/interp/auto.py#L144
     (match(r"\[.*\]", leaf.fTitle) !== nothing) && return Nooffsetjagg
     leaf isa TLeafElement && leaf.fLenType==0 && return Offsetjagg
+    !hasproperty(branch, :fClassName) && return Nojagg
 
     try
         streamer = streamerfor(f, branch.fClassName).streamer.fElements.elements[1]