When reading files where the performance is not dominated by interpretation, ~20% of the time is in the `fields...` unpacking. In this file, there are ~175 events per basket, probably because there are many branches.
```julia
julia> t.pass_dxysig
57667-element LazyBranch{Bool, UnROOT.Nojagg, Vector{Bool}}:
 0
 ⋮

julia> UnROOT.numbaskets(t.pass_dxysig.b)
328
```
After fixing the 20% `fields...` unpacking time, 75% of the remaining time is spent in `auto_T_JaggT`. See lower in this thread for the explanation.

**before**
```julia
julia> @benchmark sum(t.pass_dxysig)
BenchmarkTools.Trial: 83 samples with 1 evaluation.
 Range (min … max):  57.313 ms … 70.517 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     60.262 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   60.802 ms ±  2.639 ms  ┊ GC (mean ± σ):  0.40% ± 1.66%

 Memory estimate: 2.59 MiB, allocs estimate: 27864.
```

**after**
```julia
julia> @benchmark sum(t.pass_dxysig)
BenchmarkTools.Trial: 378 samples with 1 evaluation.
 Range (min … max):  11.502 ms … 20.477 ms  ┊ GC (min … max): 0.00% … 25.99%
 Time  (median):     12.705 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   13.239 ms ±  1.600 ms  ┊ GC (mean ± σ):  1.19% ±  4.54%

 Memory estimate: 1.62 MiB, allocs estimate: 16056.
```