@Moelf 2021-09-26
>there are ~175 events per basket

this is depressing :(

>probably because there are many branches

isn't each branch basically independent?