@aminnj 2021-09-26
I can't remember back to when or where the file was created, but I'm guessing there was some heuristic to decide basket size based on event content (jagged/nonjagged) and number of total branches in the file