@aminnj 2021-09-26
If you know how to speed this up, we can get rid of ~100% of the `auto_T_JaggT` time (for this example, at least):
![image](https://user-images.githubusercontent.com/5760027/134790061-b0dc95e9-2823-4321-8b0a-4edd1eb03c5a.png)