@aminnj 2021-09-26
Most of the `auto_T_JaggT` is spent in `JaggType` where `branch.fClassName` fails out of the `try-catch` for basic branch types.

https://github.com/tamasgal/UnROOT.jl/blob/db8938009f62f12d994433b81efe5d81eb91708f/src/utils.jl#L64-L70

oddly enough, putting a `!hasproperty(branch, :fClassName) && return Nojagg` before the `try` block eliminates the slowdown.

**before**
```julia
julia> @time for s in propertynames(t); _ = getproperty(t, s)[:]; end
  4.425899 seconds (1.67 M allocations: 242.690 MiB, 0.54% gc time)
```
**after**
```julia
julia> @time for s in propertynames(t); _ = getproperty(t, s)[:]; end
  1.554661 seconds (1.71 M allocations: 245.022 MiB, 1.59% gc time, 6.71% compilation time)
```
(yes dumb example of looping through all branches, but it can happen in the wild ;) )