@aminnj 2021-09-26
Another note, the `readtype()`s seem to collectively give a slowdown (but still tiny) around here:
https://github.com/tamasgal/UnROOT.jl/blob/db8938009f62f12d994433b81efe5d81eb91708f/src/types.jl#L70-L84
I found we can make `unpack(..., TBasketKey)` about 40% faster by adding 3 lines after L70 once we know the keylen:
```julia
bytes = read(io, fKeylen - 16)
start = -16
io = IOBuffer(bytes)
```
which accounts for the fact that we already read 16 bytes, then reads the remaining key's bytes in one go and creates a new IOBuffer. The fact that this is faster than the original IOBuffer means that julia saves overhead from asking the disk (cache) for those bytes one by one?

I didn't include it as part of the PR because it might be overkill for the performance it adds.