@tamasgal 2021-09-26
That's a totally valid improvement, go for it! :) I recall that I was considering such micro optimisations a few times but hold them back while I was figuring out how things work, just to make it easier to read and adjust the logic. You'll likely find a few other hidden gems like this 😅